fix: avoid os remove errors interruption

+++ eva-context-adaptation/gleason_context_eval.py
@@ -213,7 +213,8 @@ class GleasonContextDataset(Dataset):
         ])
 
         # Clean up
-        os.remove(tmp_csv_path)
+        if os.path.exists(tmp_csv_path):
+            os.remove(tmp_csv_path)
     
     def __len__(self) -> int:
         return len(self.meta_df)
@@ -239,7 +240,8 @@ class GleasonContextDataset(Dataset):
             img = Image.open(tmp_patch).convert("RGB")
             patch_tensor = self.transform(img)
 
-            os.remove(tmp_patch)
+            if os.path.exists(tmp_patch):
+                os.remove(tmp_patch)
 
             return patch_tensor, label, idx
 
@@ -321,7 +323,8 @@ class GleasonContextDataset(Dataset):
             region_tensor = torch.stack(tiles, dim=0)
 
             # Clean up
-            os.remove(tmp_tma)
+            if os.path.exists(tmp_tma):
+                os.remove(tmp_tma)
 
             return region_tensor, target_index, label, idx
